requirements.txt
pillow
keyboard
clipboard
pyinstaller